Add failing tests for simple token support
